Allow the list of symbols to be empty.
This commit weakens the voluptuous specification of a valid list of
symbols to allow the list to be empty.  Running viewer on the output
of the Rust model checker currently tries to run ctags on the Rust
code and (for obvious reasons) finds no C symbols to add to the symbol
table.