Don't link error traces to library source files
Viewer annotates only source files under SRCDIR, links steps of an
error trace only to lines of code in files under SRCDIR, and uses path
names relative to SRCDIR for files under SRCDIR.  With RMC, we see for
the first time errors triggered by lines of code in the RMC
implementation (eg, the Rust standard library) not under SRCDIR, and
viewer fails when it tries to generate a link to the line of code
triggering the error (in a library source file given by an absolute
path name).

With this commit, viewer will no longer try to link to a file specified
with an absolute path name (to a file not under SRCDIR).