Simplify symbol table parsing to work for both C and Rust
This commit has the effect of doubling the number of C symbols since
the list now includes fully-qualified symbols like function::1::x that
denotes local variable x in scope 1 of function.  But this doesn't
change the html generated by viewer for C code.

Proposing this pull request in favor of
https://github.com/awslabs/aws-viewer-for-cbmc/pull/42