Allow markup functions to return None when text to markup is None.
The source location for a global variable (a static variable) will
give a file name and a line number, but the function name will be
omitted (because there is no enclosing function to name).  Trying to
link a global variable to its source location will invoke markup
functions with text set to None (the missing function name).