Refactor result parsing
The prior version of the result parsing (especially the text parsing)
involved large loops that were hard to understand.  This commit
refactors the code into small methods to parse for the program
version, to parse for status messages, to parse for warnings, etc.
The result makes it easier to read, but also makes it easier to see
that the text, json, and xml parsing are producing the same results.