Refactor srcloc: Add function constructing srclocs relative to srcdir
Viewer transforms cbmc srclocs into viewer srclocs with file names
that are relative to the source root instead of relative to the
working directory.  This patch refactors the srcloc module to make
this transformation available to the coverage module that must
performs a similar transformation on the coverage descriptions
appearing in coverage data.