Repair coverage: Use correct srclocs for basic block locations
This patch repairs a bug in the parser of a basic block's coverage
description.  A basic block's coverage description lists the source
locations that contribute to the basic block. Relative path names in
these source locations are relative to the working directory, but we
want source locations that are relative to the source root.  The
current implementation is using the same path relative to the source
root for all locations.  This fails when the basic block includes
inlined code from several different files.  This patch fixes this
mistake by threading the srcdir and wkdir down to the parser of the
basic block's coverage description.