Apply visible_steps to individual text traces
A prior commit added a visible_steps generator to modify which steps
in a trace are considered visible and to omit the remaining hidden
steps from the trace.  That commit incorrectly applied the generator
to the dict mapping trace names to traces, rather than to the traces
themselves.  This commit repairs that error.