Make development mode install into a virtual environment
In the transition from dynamic packaging setup.py to static packaging
pyproject.toml, the ability to install in development mode was lost.
Installing in development mode has always been a problem with
permissions and trying not to install into system directories.
Installing in development mode required setup.py before pip 21.3, and
one of the proposed workarounds was to install development mode into a
virtual environment.  We should have been doing that from the start!

This patch rewrites 'make develop' to create a virtual environment in
VENV=/tmp/cbmc-viewer, and to install in editable mode into that
virtual environment.  It is also careful to update to pip 21.3 since
Ubuntu 20 defaults to pip 20.  Now 'make undevelop' just deletes
$VENV.