Remove xml string patching
It used to be that messages printed by cbmc before the coverage data
we care about quoted from the code, and the quoted null character '\0'
appeared as '&#0', which is an illegal character in xml, and caused
xml parsing to throw an exception.  CBMC no longer seems to insert
these illegal characters.  This patch deletes old string patching
that dealt with this.