Raise UserWarning on failure to load xml/json
Viewer used to log its failure to load cbmc's xml/json output, and
continue as best it could to produce a useful report with what it
could load.  This led people (not looking at the log) to think there
were no issues when there were.  This patch removes all special
handling of load/parse failures and just raises a UserWarning.