Modify regression script difference.py to select only cbmc-viewer jobs
Strengthen the filter used by the regression script difference.py to
select the cbmc-viewer litani jobs.  Before this commit, the script
selected all jobs in the "report" pipeline.  This commit strengthens
the filter used by difference.py to select only jobs in the "report"
pipeline with "cbmc-viewer" in the job command.