Lower 'no cbmc-viewer.json' logging from error to info
Viewer currently logs an error if the configuration file
cbmc-viewer.json is missing.  Viewer does not depend on the
configuration file, it is merely guided to better results by the
configuration file.  This change lowers the error message to an info
message (that is not displayed without the --verbose or --debug
flags).