Use cElementTree to reduce memory consumption
Parsing CBMC's results in a memory consumption that is 100 times the
size of the input file. Switching to cElementTree appears to halve the
memory footprint, which can make the difference as to whether this fits
into host memory. We could reduce the memory footprint a lot by
iterating (and cleaning up) over the input XML nodes, but this requires
a much larger re-architecting of the source.